去警告

+++ DWZShareSDKDemo/DWZShareSDKDemo/Core/DWZImageHandle.m
@@ -49,7 +49,7 @@
     // Fill and end the transparency layer
     
     
-    const float* colors = CGColorGetComponents( color.CGColor );
+    const CGFloat* colors = CGColorGetComponents( color.CGColor );
     CGContextSetRGBFillColor(c, colors[0], colors[1], colors[2], .75);
     
     contextRect.size.height = -contextRect.size.height;
